Refactored to single responsibility components
