+++ src/component/common/ConstraintAccordion/helpers.ts
